Time: 8 ms (46.15%) | Memory: 45.7 MB (28.85%) - LeetSync

+++ 1647-can-convert-string-in-k-moves/can-convert-string-in-k-moves.java
@@ -38,11 +38,8 @@ class Solution {
 
         for(int i=0; i<s.length(); i++){
             if(s.charAt(i)!= t.charAt(i)){
-                int dif =(int) (t.charAt(i) - s.charAt(i));
-
+                int dif =t.charAt(i) - s.charAt(i);
                 dif = (dif+26)%26;
-    
-                
                 map[dif]++;
             }
         }
